+++ data/src/main/java/com/cheocharm/data/repository/LoginRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.cheocharm.data.repository
 
 import com.cheocharm.data.error.ErrorData
 import com.cheocharm.data.error.toDomain
-import com.cheocharm.data.source.LoginRemoteDataSource
+import com.cheocharm.data.remote.source.LoginRemoteDataSource
 import com.cheocharm.domain.model.GoogleSignUpRequest
 import com.cheocharm.domain.model.Token
 import com.cheocharm.domain.model.MapZSignInRequest
