+++ data/src/main/java/com/cheocharm/data/repository/MyGroupsRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.cheocharm.data.repository
 import com.cheocharm.data.error.ErrorData
 import com.cheocharm.data.error.toDomain
 import com.cheocharm.data.mapper.toDomain
-import com.cheocharm.data.source.MyGroupsRemoteDataSource
+import com.cheocharm.data.remote.source.MyGroupsRemoteDataSource
 import com.cheocharm.domain.model.Group
 import com.cheocharm.domain.repository.MyGroupsRepository
 import javax.inject.Inject
