+++ remote/src/main/java/com/cheocharm/remote/source/WriteRemoteDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.cheocharm.remote.source
 
 import com.cheocharm.data.error.ErrorData
-import com.cheocharm.data.source.WriteRemoteDataSource
+import com.cheocharm.data.remote.source.WriteRemoteDataSource
 import com.cheocharm.domain.model.TempDiary
 import com.cheocharm.domain.model.WriteDiaryRequest
 import com.cheocharm.domain.model.WriteImageRequest
